Goldstone object fetching is now working (but we are saving too many revisions of objects):
ingest/management/commands/fetch_goldstone_targets.py: Update log message.
ingest/models.py: Change the __unicode__ repr so we return the name if that is not blank and the provisional desigination is blank.
ingest/sources_subs.py: Strip off any leading or trailing space and replace internal space with a plus sign in fetch_mpcorbit(). Trap for no element tables in HTML page. Handle the case of Goldstone objects with a specific date and a name. Trap the case where we read off the end of the table and no longer get a year in the first column.
ingest/tests.py: Add tests for the object id parser in parse_goldstone_chunks().
ingest/views.py: Pass through dbg flag setting.