Relax validation on window start otherwise it's impossible to schedule for the current day:
core/forms.py: Relax validation on window start to check it's not more a day in the past. Otherwise it's impossible to schedule for the current day/night if the site is observing. I'm not very happy with this fix - a better one would to check if window_end - slot_length < current UTC time but the current clean_start doesn't seem to have access to these values.
core/templates/core/schedule_confirm.html: Display window in schedule confirm output. Somewhat ugly, but very helpful right now.