Cherry-pick Sarah's fix for the ZP determination
Added adjustable tolerance for the zeropoint standard deviation to be able to adjust the tightness of the cross-match. Fixed 'nan' value for zeropoint standard deviation when only 1 star in cross-match by requiring at least 2 stars to be in the cross-match. If either one star or no stars are in the cross-match, the zeropoint and standard deviation will report a dummy value (40.0 and 10.0, respectively) and the header values will not be updated. If 2 or more stars are in the cross-match, the calculated zeropoint and standard deviation will be reported and the header values will be updated. Fixed issue with multiple Frames being created each time a new zeropoint is calculated and updated in the header.