Further integration of candidate display:
Add a button to take you to the candidates views if the Block is observed and has Candidates.
Add separation to the columns in the template (this may need a better way of doing the ephemeris computation once and then computing the separation for each candidate in the block).
Reduce the number of output digits in the template.