Migrate to modern operating system configuration step
The modern way to add operating system configuration to a Docker
container is to build a directory which overlays the filesystem root
with all of the configuration files. This directory is copied into the
container in a single step.

None of the configuration files are modified, they are only moved into a
new location.