Update dependencies, removing last need for rise_set
Since the need for reqdbclient has gone and it has been removed from the requirements, we weren't getting the implicit dependencies on rise_set and pyslalib. Support for the old style of orbital elements has been removed from compute_ephem(), removing the last need on the Angle part of rise_set.
Add in explicit requirement for pyslalib (which we do need).