Updates to block status logic and update_blocks to try and prevent data not being ingested.
Increase logging and add extra logging messages to aid diagnosis.
Increase time from 120 to 240 minutes before we declare a block inconsistent and mark it active to work round data transfer/pipeline/archive delays.
Refactor logic in block_status:
If we have images for the request, we will ingest them no matter how many we have (previously needed 3).
If we have images for the request, we update block.when_observed for each image.
The block is only set inactive if end time has passed AND all images have been received.
The block.num_observed is only set to >=1 if we got 3 or more images back (less than 3 is not useful for astrometry of asteroids and so doesn't count as "observed" for our purposes)