Fix cross-identification of NEO candidates to numbered asteroids.
Note: the comet number is now correctly truncated (i.e. '46P' not '0046P') which is used as the name being used to search for the matching Body. In order to avoid duplicate objects, the 2 Body's for 46P and 60P in the DBs will also need truncating before running the fixed update_crossids command.