Use supercronic as cron daemon
The supercronic cron daemon is a much better alternative cron daemon for
Docker (and Kubernetes) use. Since this cron daemon is designed to work
in a Docker environment, it has the following useful features, which let
us remove a bunch of workarounds from this code:

- Does not clear environment variables before running a task (no need
  for cronwrapper workaround).
- Logging goes to stdout/stderr by default (no need to run in a special
  debug mode).

The supercronic daemon will be started by Kubernetes as a second
container within the Pod.