Filter out sources with -ve flux from the table before converting to magnitudes.
This prevents the Runtime warning that appeared after we changed to converting the whole column at once and stops the get_or_create_CatalogSources() throwing odd "Unknown column 'nane0' in 'field list'" errors. It is however a slight loss of functionality compared to get_catalog_items_old() which returned, and allowed storage of, the original -ve flux value for marginally detected sources.