More additions for adding MRO support:
Add mro_ldac_catalog_mapping() routine.
Modify swop_ldac_catalog_mapping when adding above to allow passing the fixed_vales_map and updating it with observatory-specific values.
Only multiply the pixelscale by 3600 if it's a celestial WCS and not a pixel-based "null" WCS.
Only test the CD matrix in the compare_headers() test helper if it's a celestial WCS.