Various ADES updates:
Add a 'ades=' option to Frame.map_filter for mapping to ADES-acceptable
band values (We're not actually using this right now as the mapping is
actually a much more complicated function of the filter and the
photometric catalog)
Add 'high_precision_mag' mode for ADES PSV lines and headers.
logSNR is actually a DecimalTypeW6 so only allows for 5 characters plus
1 for optional sign. Left justify in it's field for the (usual) case
  where there isn't a sign in front (for SNR>1)
Update tests for logSNR change; add 1 for high precision mag mode.